@@ -18,28 +18,28 @@
         }
     },
     "require": {
-        "php": ">=5.3.23",
+        "php": ">=5.5",
         "bshaffer/oauth2-server-php": "~1.7",
-        "zendframework/zend-crypt": ">=2.3,<2.5",
-        "zendframework/zend-http": ">=2.3,<2.5",
-        "zendframework/zend-mvc": ">=2.3,<2.5",
-        "zendframework/zend-servicemanager": ">=2.3,<2.5",
+        "zendframework/zend-crypt": "~2.3",
+        "zendframework/zend-http": "~2.3",
+        "zendframework/zend-mvc": "~2.3",
+        "zendframework/zend-servicemanager": "~2.3",
         "zfcampus/zf-api-problem": "~1.0",
         "zfcampus/zf-content-negotiation": "~1.0"
     },
     "require-dev": {
         "phpunit/phpunit": "~4.7",
-        "squizlabs/php_codesniffer": "^2.3",
-        "zendframework/zend-config": ">=2.3,<2.5",
-        "zendframework/zend-form": ">=2.3,<2.5",
-        "zendframework/zend-loader": ">=2.3,<2.5",
-        "zendframework/zend-log": ">=2.3,<2.5",
-        "zendframework/zend-modulemanager": ">=2.3,<2.5",
-        "zendframework/zend-serializer": ">=2.3,<2.5",
-        "zendframework/zend-test": ">=2.3,<2.5",
-        "zendframework/zend-authentication": ">=2.3,<2.5",
-        "zendframework/zend-i18n": ">=2.3,<2.5",
-        "zendframework/zend-db": ">=2.3,<2.5",
+        "squizlabs/php_codesniffer": "^2.3.1",
+        "zendframework/zend-config": "~2.3",
+        "zendframework/zend-form": "~2.3",
+        "zendframework/zend-loader": "~2.3",
+        "zendframework/zend-log": "~2.3",
+        "zendframework/zend-modulemanager": "~2.3",
+        "zendframework/zend-serializer": "~2.3",
+        "zendframework/zend-test": "~2.3",
+        "zendframework/zend-authentication": "~2.3",
+        "zendframework/zend-i18n": "~2.3",
+        "zendframework/zend-db": "~2.3",
         "mockery/mockery": "0.9.*"
     },
     "autoload": {
