@@ -30,18 +30,18 @@ class PdoAdapterFactory implements FactoryInterface
 
         $username = isset($config['zf-oauth2']['db']['username']) ? $config['zf-oauth2']['db']['username'] : null;
         $password = isset($config['zf-oauth2']['db']['password']) ? $config['zf-oauth2']['db']['password'] : null;
-        $options  = isset($config['zf-oauth2']['db']['options']) ? $config['zf-oauth2']['db']['options'] : array();
+        $options  = isset($config['zf-oauth2']['db']['options']) ? $config['zf-oauth2']['db']['options'] : [];
 
-        $oauth2ServerConfig = array();
+        $oauth2ServerConfig = [];
         if (isset($config['zf-oauth2']['storage_settings']) && is_array($config['zf-oauth2']['storage_settings'])) {
             $oauth2ServerConfig = $config['zf-oauth2']['storage_settings'];
         }
 
-        return new PdoAdapter(array(
+        return new PdoAdapter([
             'dsn'      => $config['zf-oauth2']['db']['dsn'],
             'username' => $username,
             'password' => $password,
             'options'  => $options,
-        ), $oauth2ServerConfig);
+        ], $oauth2ServerConfig);
     }
 }
