@@ -31,12 +31,12 @@ class AuthControllerWithMongoAdapterTest extends AbstractHttpControllerTestCase
             $this->markTestSkipped($e->getMessage());
         }
         $this->db = $client->selectDB('zf_oauth2_test');
-        $this->db->oauth_clients->insert(array(
+        $this->db->oauth_clients->insert([
             'client_id'     => 'testclient',
             'client_secret' => '$2y$14$f3qml4G2hG6sxM26VMq.geDYbsS089IBtVJ7DlD05BoViS9PFykE2',
             'redirect_uri'  => '/oauth/receivecode',
             'grant_types'   => null
-        ));
+        ]);
 
         $this->getApplicationServiceLocator()->setService('MongoDB', $this->db);
     }
@@ -90,14 +90,14 @@ class AuthControllerWithMongoAdapterTest extends AbstractHttpControllerTestCase
 
     public function testAuthorizeCode()
     {
-        $queryData = array(
+        $queryData = [
             'response_type' => 'code',
             'client_id'     => 'testclient',
             'state'         => 'xyz',
             'redirect_uri'  => '/oauth/receivecode'
-        );
+        ];
 
-        $this->dispatch('/oauth/authorize?' . http_build_query($queryData), 'POST', array('authorized' => 'yes'));
+        $this->dispatch('/oauth/authorize?' . http_build_query($queryData), 'POST', ['authorized' => 'yes']);
         $this->assertTrue($this->getResponse()->isRedirect());
         $this->assertControllerName('ZF\OAuth2\Controller\Auth');
         $this->assertActionName('authorize');
